// src/richtexteditor/src/common/interface.ts
export interface IUndoCallBack {
    callBack?: Function
    event?: Object
}